refactor: Remove persist WriteTransformer
Turns out that we can just write the chunks since we computed the hashes
in an earlier step.

Persist ComputeTransformer now takes over some of the work of
WriteTransformer.

Towards #671
